backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.9
anthropic==0.18.1
openai==1.12.0
supabase==2.3.4
pydantic==2.6.0
pydantic-settings==2.1.0
python-dotenv==1.0.1
pypdf2==3.0.1
python-docx==1.1.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9

